requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-decouple==3.8
pydantic[email]==2.6.4
python-multipart==0.0.9
Jinja2==3.1.3
aiofiles==23.2.1
